requirements.txt
face_recognition==1.3.5
opencv-python==4.8.1.78
numpy==1.24.3
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
python-dotenv==1.0.0
